added support for 80000ish sized texcoloffsets, now in 0x5000 and 0x6000 regions. some memory fields moved around. 0x7000 scratch page support. this has made doom2 loading possible, but currently some bugs are preventing level loading.

+++ g_setup.c
@@ -89,17 +89,16 @@ extern int16_t             numtextures;
 //
 uint16_t     R_CheckTextureNumForName(int8_t *name)
 {
-	uint8_t         i;
+	uint16_t         i;
 	texture_t __far* texture;
 	int8_t texturename[8];
 	// "NoTexture" marker.
 	if (name[0] == '-')
 		return 0;
 
-
 	for (i = 0; i < numtextures; i++) {
 		texture = (texture_t __far*)&(texturedefs_bytes[texturedefs_offset[i]]);
-			
+
 		FAR_strncpy(texturename, texture->name, 8);
 		//DEBUG_PRINT("\n %.8Fs %8s %8s %i %Fp", texture->name, texturename, name, texture->name);
